requirements.txt
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Django==2.2
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
python-jwt==3.3.0
django-filter==2.4.0